Refresh owner evidence and split projection tests for size policy
The regenerated tutorial assets and the execution-manifest update
changed files pinned by scenario-owner-evidence, and form_ops_tests
grew past the 1000-line hard limit. Recompute the artifact digests
and move the pure effective_agent_type_ids projection test into its
own form_projection_tests module beside the code it covers.
